Updated solution to new name.
