+++ IvanStoychev.StringExtensions.Tests/SelectorTests.cs
@@ -1,7 +1,7 @@
 ﻿using Xunit;
 using System;
 
-namespace Useful.String.Extensions.Tests
+namespace IvanStoychev.StringExtensions.Tests
 {
     public class SelectorTests
     {
