+++ IvanStoychev.StringExtensions.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29519.181
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IvanStoychev.StringExtensions", "IvanStoychev.StringExtensions\IvanStoychev.StringExtensions.csproj", "{412B295D-E36E-475C-9BF5-703C066A5D22}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IvanStoychev.StringExtensions.Tests", "IvanStoychev.StringExtensions.Tests\IvanStoychev.StringExtensions.Tests.csproj", "{F4B4D2C9-3B20-4987-BB02-72811B04C0DF}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{412B295D-E36E-475C-9BF5-703C066A5D22}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{412B295D-E36E-475C-9BF5-703C066A5D22}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{412B295D-E36E-475C-9BF5-703C066A5D22}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{412B295D-E36E-475C-9BF5-703C066A5D22}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F4B4D2C9-3B20-4987-BB02-72811B04C0DF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F4B4D2C9-3B20-4987-BB02-72811B04C0DF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F4B4D2C9-3B20-4987-BB02-72811B04C0DF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F4B4D2C9-3B20-4987-BB02-72811B04C0DF}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {63C36473-380A-41B2-9B71-37D6DABC496B}
+	EndGlobalSection
+EndGlobal
