--- Useful.String.Extensions.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29519.181
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Useful.String.Extensions", "Useful.String.Extensions\Useful.String.Extensions.csproj", "{54A33500-5915-4FD2-A5D2-617EDD77F461}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Useful.String.Extensions.Tests", "Useful.String.Extensions.Tests\Useful.String.Extensions.Tests.csproj", "{01E89E00-6896-420C-88C5-02C451109D4A}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{54A33500-5915-4FD2-A5D2-617EDD77F461}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{54A33500-5915-4FD2-A5D2-617EDD77F461}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{54A33500-5915-4FD2-A5D2-617EDD77F461}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{54A33500-5915-4FD2-A5D2-617EDD77F461}.Release|Any CPU.Build.0 = Release|Any CPU
-		{01E89E00-6896-420C-88C5-02C451109D4A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{01E89E00-6896-420C-88C5-02C451109D4A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{01E89E00-6896-420C-88C5-02C451109D4A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{01E89E00-6896-420C-88C5-02C451109D4A}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {63C36473-380A-41B2-9B71-37D6DABC496B}
-	EndGlobalSection
-EndGlobal
