--- Useful.String.Extensions/Enums.cs
@@ -1,10 +1,0 @@
-﻿namespace Useful.String.Extensions
-{
-    public enum StringInclusionOptions
-    {
-        IncludeNone,
-        IncludeStart,
-        IncludeEnd,
-        IncludeAll
-    }
-}
