--- Useful.String.Extensions/Useful.String.Extensions.csproj
@@ -1,39 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.0</TargetFramework>
-    <Authors>Ivan Stoychev</Authors>
-    <Title>Useful String Extensions</Title>
-    <Language>English</Language>
-    <Description>A .Net Core class library of useful extension methods for the "string" class in C#.</Description>
-    <Summary>A .Net Core class library collection of useful extension methods for the "string" class in C#.</Summary>
-    <PackageProjectUrl>https://github.com/IvanStoychev/Useful.String.Extensions/wiki</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/IvanStoychev/Useful.String.Extensions</RepositoryUrl>
-    <PackageTags>c#, string, useful, extension</PackageTags>
-    <RepositoryType>git</RepositoryType>
-    <PackageIcon>strong.png</PackageIcon>
-    <Copyright>Copyright (c) 2020 Ivan Stoychev. All rights reserved.</Copyright>
-    <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
-    <NeutralLanguage>English</NeutralLanguage>
-    <PackageReleaseNotes>- Added "SubstringEnd" methods, which select a substring ending at a given string.
-- Added powerful "Substring" methods, which allow you to specify which occurrence of a start and end substring to use as the start and end of an extraction and what comparison rules to use.</PackageReleaseNotes>
-    <Version>1.5.0</Version>
-    <PackageLicenseFile>LICENSE</PackageLicenseFile>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
-    <DocumentationFile>J:\Users\Administrator\Source\Repos\Useful.String.Extensions\Useful.String.Extensions\bin\Release\netcoreapp3.0\Useful.String.Extensions.xml</DocumentationFile>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Include="..\LICENSE">
-      <Pack>True</Pack>
-      <PackagePath></PackagePath>
-    </None>
-    <None Include="strong.png">
-      <Pack>True</Pack>
-      <PackagePath></PackagePath>
-    </None>
-  </ItemGroup>
-
-</Project>
